+++ functions/download.js
@@ -1,24 +1,25 @@
+// core modules
+import { writeFile } from "node:fs/promises";
+
 // colors module
 import colors from "colors";
 
 // url api
 import { wallpapersURL } from "../api/wallpapersURL.js";
 
-// save file
-import { stackSave } from "../utils.js";
-
 /**
   * sol, moon wallpapers downloader
   * @async
-  * @param {"sol-moon" | "dimensions" | "seyyahi"} opt
+  * @param {"sol-moon" | "dimensions" | "seyyahi2" | "ancient-mistery" | "tsuky-no-emily"} opt
   * @param {string} filename
   * @returns {Promise<void>}
 */
-export const wallpaperDownload = async (opt, filename) => {	
+export const wallpaperDownload = async (opt, filename) => {
   try {
     const { data } = await wallpapersURL.get(`/${opt}/download/${filename}`);
-    stackSave(filename, data);
-  } catch(err) {
+
+    writeFile(filename, data, {encoding: "base64"});
+  } catch (err) {
     console.error(colors.red(err.message));
   }
 };